* origin/staging:
  Snapshot of v1.5.0gh 2019-06-28 (Internal repo commit 25be93d) v1.5.0rc6
  Snapshot of v1.5.0gh 2019-05-15 (Internal repo commit 1b6fc4e) v1.5.0rc5
  Snapshot of v1.5.0gh 2019-05-13 (Internal repo commit 3f374fc)
  Snapshot of v1.5.0gh 2019-05-01 (Internal repo commit 0693997) v1.5.0rc4
  Snapshot of develop 2019-04-18 (v1.5.0rc2-ish) (Internal repo commit 62d86d4)
  Snapshot of develop 2019-04-08 (v1.5.0rc0) (Internal repo commit 29fb19e)
  Snapshot of develop 2019-03-22 (Internal repo commit 0f53b3e)
  Snapshot of develop 2019-03-08 (Internal repo commit 175deb8)
  Snapshot of develop 2019-02-19 (Internal repo commit c1adee9)
  Snapshot of develop 2019-01-28 (Internal repo commit 088bff2)
  Snapshot of develop 2019-01-18 (Internal repo commit c59cbad)
  Snapshot of develop 2018-12-20 (Internal repo commit 159e985)
  Snapshot of develop 2018-12-07 (Internal repo commit 59a30f0)
  Snapshot of develop 2018-11-30 (Internal repo commit 9cd5cb0)
  Snapshot of develop 2018-11-16(b) (Internal repo commit 856a006)
  Snapshot of develop 2018-11-16 (Internal repo commit f5a65a4)
  Snapshot of develop 2018-11-09 (Internal repo commit 49e6857)
  Snapshot of develop 2018-11-02 (Internal repo commit b882288)
  Snapshot of develop 2018-10-19 (Internal repo commit 959fcfb)
  Snapshot of develop 2018-08-17 (Internal repo commit dc49e90)
  Snapshot of develop 2018-08-03 (Internal repo commit 27f7cce)
  Snapshot of develop 2018-06-29 (Internal repo commit 341c586)
  Snapshot of develop 2018-06-15 (Internal repo commit 3e8abb0)
  Fix install issue AV-4255
  Snapshot of develop 2018-05-24 (Internal repo commit 3674b3b)
  Snapshot of develop 2018-05-10 (Internal repo commit abc775e)
  Snapshot of develop 2081-04-27 (Internal repo commit 9c3b3be)
  Snapshot of develop 2018-04-16 (Internal repo commit 7612d54)
  Snapshot of develop 2018-04-10 (Internal repo commit 99abb93)
  Snapshot of develop 2018-04-10 (Internal repo commit a0a8951)

+++ runtime/dataplane/transport/impl/include/DtHandshakeControl.h
@@ -47,7 +47,7 @@ namespace OCPI {
   // Token definition
   typedef int32_t ControlToken;
   typedef int32_t BooleanToken;
-  typedef int32_t PortId; // signed to allow a sentinel of -1
+  typedef uint32_t PortId; // was signed to allow a sentinel of -1, but not used that way anymore
 
   /**********************************
    *  SMB communications structures
@@ -240,10 +240,10 @@ namespace OCPI {
     assert(length <= maxXferLength);
     return (uint32_t)
       ((length & ~(UINT32_MAX << lengthBits)) | // length is LSB
-       ((eof ? 1 : 0) << lengthBits) |          // EOF independent of length, above length
+       ((eof ? 1u : 0) << lengthBits) |          // EOF independent of length, above length
        (1 << (lengthBits+1)) |                  // always-1 is above eof
                                                 // truncation indicator is above always-1
-       (opcode << (32 - opCodeBits)));          // high byte is opcode
+       ((uint32_t)opcode << (32 - opCodeBits)));          // high byte is opcode
   }
   inline void unpackXferMetaData(uint32_t md, size_t &length, uint8_t &opcode, bool &eof,
 				 bool &truncate) {
